backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
opencv-python

numpy
pydantic
python-dotenv




tqdm
